Catch all Verto errors when converting
